Update and simplify script
After moving to Arch I decided to give this forgotten script a try. It didn't work. It was also very complicated. URL parsing from JSON is a crazy thing to try to do in a bash script. I don't have a Mac to test these changes on but hopefully everything still works. There were a number of overlapping PRs over the years that made it into this revision. I've tried to add credit to the original PRs where applicable.

Summary of changes:

1. Merge latest and boost code paths (From @Lyncredible in #39)
2. Use https for all requests if -s (From @Lyncredible in #39)
3. Update URL patterns  (From @Lyncredible in #39, @microdog in #36, @nitrogear in #23)
4. Add UHD support (fixes #43)
5. Remove contributing.json as it wasn't used anymore